requirements.txt
click==8.1.3
Flask==2.2.2
Flask-WTF==1.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
pytz-deprecation-shim==0.1.0.post0
rel==0.4.7
tzdata==2022.7
tzlocal==4.2
websocket-client==1.4.2
Werkzeug==2.2.2
WTForms==3.0.1
